[#270] Remove iox2 and merge it with node
