+++ examples/rust/complex_data_types/complex_data_types.rs
@@ -54,7 +54,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let subscriber = service.subscriber_builder().create()?;
     let mut counter = 0;
 
-    while let Iox2Event::Tick = Iox2::wait(CYCLE_TIME) {
+    while let NodeEvent::Tick = node.wait(CYCLE_TIME) {
         // ComplexDataType as a size of over 30MB, we need to perform a placement new
         // otherwise we will encounter a stack overflow in debug builds.
         // Therefore, we acquire an uninitialized sample, use the PlacementDefault
