+++ iceoryx2-ffi/ffi/src/publisher.rs
@@ -53,7 +53,7 @@ pub extern "C" fn run_publisher(seconds: u32) -> i32 {
 
     let mut remaining_seconds = seconds;
 
-    while let Iox2Event::Tick = Iox2::wait(CYCLE_TIME) {
+    while let NodeEvent::Tick = node.wait(CYCLE_TIME) {
         counter += 1;
         let sample = publisher.loan_uninit();
