+++ iceoryx2/src/lib.rs
@@ -289,9 +289,6 @@ mod compiletests;
 /// Handles iceoryx2s global configuration
 pub mod config;
 
-/// Central instance that handles all incoming events, the event loop
-pub mod iox2;
-
 /// Central instance that owns all service entities and can handle incoming event in an event loop
 pub mod node;
