+++ iceoryx2/src/prelude.rs
@@ -11,9 +11,7 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
 pub use crate::config::Config;
-pub use crate::iox2::Iox2;
-pub use crate::iox2::Iox2Event;
-pub use crate::node::{node_name::NodeName, Node, NodeBuilder, NodeState};
+pub use crate::node::{node_name::NodeName, Node, NodeBuilder, NodeEvent, NodeState};
 pub use crate::port::event_id::EventId;
 pub use crate::service::messaging_pattern::MessagingPattern;
 pub use crate::service::{
